+++ luscii_patient_actions_sdk_android/lib/luscii_patient_actions_sdk_android.dart
@@ -24,9 +24,26 @@ class LusciiPatientActionsSdkAndroid extends LusciiPatientActionsSdkPlatform {
   }
 
   @override
-  Future<void> initialize({bool androidDynamicTheming = false}) async {
-    await methodChannel.invokeMethod<void>('initialize', <String, bool>{
+  Future<void> initialize({
+    bool androidDynamicTheming = false,
+    LusciiEnvironment environment = LusciiEnvironment.production,
+  }) async {
+    String environmentString = 'production';
+    switch (environment) {
+      case LusciiEnvironment.production:
+        environmentString = 'production';
+        break;
+      case LusciiEnvironment.acceptance:
+        environmentString = 'accept';
+        break;
+      case LusciiEnvironment.test:
+        environmentString = 'test';
+        break;
+    }
+
+    await methodChannel.invokeMethod<void>('initialize', <String, dynamic>{
       'androidDynamicTheming': androidDynamicTheming,
+      'environment': environmentString,
     });
   }